Workspace/flatswapp/requirements.txt
Babel==2.8.0
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==2.1.5
django-google-maps==0.12.1
django-phone-field==1.8.0
django-phonenumber-field==4.0.0
idna==2.9
oauthlib==3.1.0
olefile==0.46
phonenumbers==8.11.4
Pillow==7.0.0
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
urllib3==1.25.8
wincertstore==0.2
